+++ yggdrasil_decision_forests/learner/decision_tree/generic_parameters.h
@@ -61,6 +61,7 @@ constexpr char kCategoricalAlgorithmRandom[] = "RANDOM";
 constexpr char kHParamSplitAxis[] = "split_axis";
 constexpr char kHParamSplitAxisAxisAligned[] = "AXIS_ALIGNED";
 constexpr char kHParamSplitAxisSparseOblique[] = "SPARSE_OBLIQUE";
+constexpr char kHParamSplitAxisMhldOblique[] = "MHLD_OBLIQUE";
 
 constexpr char kHParamSplitAxisSparseObliqueNumProjectionsExponent[] =
     "sparse_oblique_num_projections_exponent";
@@ -80,6 +81,11 @@ constexpr char kHParamSplitAxisSparseObliqueNormalizationNone[] = "NONE";
 constexpr char kHParamSplitAxisSparseObliqueNormalizationStandardDeviation[] =
     "STANDARD_DEVIATION";
 constexpr char kHParamSplitAxisSparseObliqueNormalizationMinMax[] = "MIN_MAX";
+
+constexpr char kHParamSplitAxisMhldObliqueMaxNumAttributes[] =
+    "mhld_oblique_max_num_attributes";
+constexpr char kHParamSplitAxisMhldObliqueSampleAttributes[] =
+    "mhld_oblique_sample_attributes";
 
 constexpr char kHParamSortingStrategy[] = "sorting_strategy";
 constexpr char kHParamSortingStrategyInNode[] = "IN_NODE";